+++ app/src/main/java/org/fitchfamily/android/symphony/Song.java
@@ -19,6 +19,14 @@
 
 package org.fitchfamily.android.symphony;
 
+import android.content.ContentUris;
+import android.content.Context;
+import android.graphics.Bitmap;
+import android.graphics.BitmapFactory;
+import android.media.MediaMetadataRetriever;
+import android.net.Uri;
+import android.util.Log;
+
 /**
  * Created by tfitch on 7/5/17.
  */
@@ -65,4 +73,24 @@ public class Song {
     public String getComposer(){return composer;}
     public int getTrack(){return track;}
 
+    public Bitmap getArtwork(Context mContext) {
+        Bitmap artwork = null;
+
+        try {
+            Uri trackUri = ContentUris.withAppendedId(
+                    android.provider.MediaStore.Audio.Media.EXTERNAL_CONTENT_URI, id);
+            MediaMetadataRetriever mmr = new MediaMetadataRetriever();
+            mmr.setDataSource(mContext, trackUri);
+
+            byte[] data = mmr.getEmbeddedPicture();
+            //coverart is an Imageview object
+
+            // convert the byte array to a bitmap
+            artwork = BitmapFactory.decodeByteArray(data, 0, data.length);
+        } catch (Exception e) {
+            Log.e("MUSIC SERVICE", "Error getting album artwork", e);
+            artwork = null;
+        }
+        return artwork;
+    }
 }